[KVConnector] Fix int32 overflow in hf3fs gather/scatter kernels
Promote token_idx to int64 at load in kv_cache_gather_kernel and
kv_cache_scatter_kernel so every `token_idx * hidden_size` offset (including the
V-half `(total_token_in_kvcache + token_idx) * hidden_size`) addresses the
offload cache correctly past 2**31 elements.

+++ vllm/distributed/kv_transfer/kv_connector/v1/hf3fs/utils/gather_scatter_helper.py
@@ -24,7 +24,7 @@ def kv_cache_scatter_kernel(
     if layer_idx >= num_layers or token_pos >= num_tokens_in_block:
         return
 
-    token_idx = tl.load(token_indices_ptr + token_pos)
+    token_idx = tl.load(token_indices_ptr + token_pos).to(tl.int64)
     kv_cache_ptr = tl.cast(tl.load(kv_cache_ptrs_ptr + layer_idx), source_ptr.dtype)
 
     if token_idx >= total_token_in_kvcache:
@@ -83,7 +83,7 @@ def kv_cache_gather_kernel(
     if layer_idx >= num_layers or token_pos >= num_tokens_in_block:
         return
 
-    token_idx = tl.load(token_indices_ptr + token_pos)
+    token_idx = tl.load(token_indices_ptr + token_pos).to(tl.int64)
     kv_cache_ptr = tl.cast(tl.load(kv_cache_ptrs_ptr + layer_idx), dst_ptr.dtype)
 
     if token_idx >= total_token_in_kvcache:
